# MAINT: avoid nested dispatch in numpy.core.shape_base