@eric-wieser 2019-06-01
I'm still -0.5 on this whole thing. 20% performance gains for very small arrays don't seem like a strong argument for me for making things harder for subclasses to hook into. The readability cost is ok here, but

> But I don't think we should consider ourselves obligated to guarantee these implementation details

I don't see any reason not to guarantee that `vstack` and `hstack` will be implemented in terms of concatenate and `atleast2d`, since that's exactly its semantics.

There are probably cases like the `mean` / `median` thing where our internals could change. Imagine a `foo(arr)` method that currently delegates to `bar(arr)`. We change it to internally call `baz(arr)` instead. One of two things happens to downstream subclasses:

* Their code continues to work, as they already implement `baz(arr)` with sensible semantics
* Their code does not work. They release an update to their library, and say _"if you want to use numpy 1.17 and newer make sure to update to our new version"_. This would happen if:
  * They do not implement `baz(arr)` at all. By making this change, we just caused a downstream library to improve and now support `baz(arr)` too.
  * They implement `baz(arr)` with incompatible semantics. They should think carefully about whether this was a good idea. Either they fix `baz(arr)` (good), or they decide they need to overload `foo(arr)` specially after all.

I suppose this all predicates on having a `__numpy_implementation__` attribute available to use _inside_ `__array_function__`.