@shoyer 2019-06-01
OK, the latest version only removes the internal uses inside `np.block`. I think there is a pretty clear case in merging this to avoid a performance regression, given that https://github.com/numpy/numpy/pull/13186 may not make it into 1.17. Also, it's just a bad idea to *partially* support subclasses in a shape dependent fashion -- we should either support subclasses always with a function or never. Value dependent semantics are unlikely to be tested properly, which vastly increases the likelihood of bugs.