@mhvk 2019-06-02
@shoyer - for `np.block`, I agree the case is better given that the path taken depends on the shape. Also, forcing a specific override means people will be ready for the C implementation.

p.s. I thought I should not let my suggested optimization of `hstack` and `vstack` slip by: see #13697. About as good as what you had by bypassing dispatch.